requirements.txt
absl-py==1.2.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cycler==0.11.0
decorator==5.1.1
efficientnet==1.1.1
flatbuffers==2.0.7
fonttools==4.37.1
gast==0.5.3
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.21.2
importlib-metadata==4.12.0
ipython==7.9.0
jedi==0.18.1
joblib==1.1.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
networkx==2.6.3
numba==0.56.0
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pooch==1.6.0
prompt-toolkit==2.0.10
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycparser==2.21
pydub==0.25.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
traitlets==5.3.0
typing-extensions==4.3.0
urllib3==1.26.12
Wave==0.0.2
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
